feat: add tests for hyperboards, reduce number of db calls (#182)
